Split the plugin into multiple modules
